Refactors Provider setup to not extend on instance
Explicitelly require inheriting from a Base encryptor to make sure it
implements the base methods instead of adding those to the instance on
runtime.
